advanced mysql

+++ 0x00-MySQL_Advanced/6-bonus.sql
@@ -1,7 +1,7 @@
 --addbonus
 DROP PROCEDURE IF EXISTS AddBonus;
 DELIMITER $$
-CREATE PROCEDURE AddBonus(user_id INT, project_name VARCHAR(255), score FLOAT)
+CREATE PROCEDURE AddBonus(user_id INT, project_name VARCHAR(255), score INT)
 BEGIN
 DECLARE project_id INT;
